Some cleanup, and made SingleActivity into a clickable activity in the list.
