+++ app/src/main/AndroidManifest.xml
@@ -20,8 +20,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".java.ASingleActivity" />
-        <activity android:name=".java.CSimpleOperatorsActivity" />
+        <activity android:name=".java.sample.SingleActivity" />
         <activity android:name=".java.sample.ButtonRapidClickActivity"/>
         <activity android:name=".java.sample.ThreadSwitchActivity"/>
         <activity android:name=".java.sample.FailingApiCallActivity"/>
@@ -30,7 +29,6 @@
         <activity android:name=".java.sample.PinActivity"/>
         <activity android:name=".java.sample.ListActivity"/>
         <activity android:name=".java.sample.ListBusActivity"/>
-
         <activity android:name=".kotlin.sample.ExtensionActivity"/>
     </application>
