+++ app/src/main/res/layout/activity_single.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".java.ASingleActivity">
+    tools:context=".java.sample.SingleActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
